+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Mapbox welcomes participation and contributions from everyone.
 * Remove `Annotation.featureIdentifier` used to connect with View Annotations, now `Annotation.id` should be used instead.
 * Rename `PuckBearingSource` to `PuckBearing` in location component plugin.
 * Remove deprecated overloaded `Style.setStyleGeoJSONSourceData(sourceId: String, data: GeoJSONSourceData)` method.
+* Rename `MapboxMap.setMemoryBudget` to `MapboxMap.setTileCacheBudget` and make it non-experimental.
 * Update Mapbox styles to latest versions.
 
 | Style             | Before                                       | After                                        |